renaming files, adding docs, adjusting imports and fixing func style
